requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
autogen-agentchat==0.5.3
autogen-core==0.5.3
autogen-ext==0.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fastapi==0.115.12
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jiter==0.9.0
jsonref==1.1.0
kiwisolver==1.4.8
matplotlib==3.10.1
networkx==3.4.2
numpy
openai==1.75.0
opentelemetry-api==1.32.1
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
protobuf==5.29.4
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
termcolor==3.0.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
zipp==3.21.0
llama-index-core>=0.10.0

# LLM v2 additional dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
qdrant-client>=1.7.0
